Fix vertex[0] bug.  Speed up calculation significantly.
For some reason reusing the same variable name as the result of a list comprehension on itself (for the vertex stride) was messing up the data.  Simply renaming to unique names fixed it.

Updated cdist to not run independently on each element.  Used the built-in euclidean distance and batch calls to take advantage of the vectorization.